refactor(parser): split parser into internal modules
